requirements.txt
flask
flask_login
flask-migrate
flask-sqlalchemy
flask_httpauth
redis
pytest
easydict
click